--- a/flecs.h
+++ b/flecs.h
@@ -1610,13 +1610,15 @@ void* flecs_bdup(
 extern "C" {
 #endif
 
-typedef uint64_t ecs_map_key_t;
+typedef uint64_t ecs_map_data_t;
+typedef ecs_map_data_t ecs_map_key_t;
+typedef ecs_map_data_t ecs_map_val_t;
 
 /* Map type */
 typedef struct ecs_bucket_entry_t {
-    struct ecs_bucket_entry_t *next;
     ecs_map_key_t key;
-    /* payload right after key. */
+    ecs_map_val_t value;
+    struct ecs_bucket_entry_t *next;
 } ecs_bucket_entry_t;
 
 typedef struct ecs_bucket_t {
@@ -1626,7 +1628,6 @@ typedef struct ecs_bucket_t {
 typedef struct ecs_map_t {
     ecs_bucket_t *buckets;
     ecs_bucket_t *buckets_end;
-    int16_t elem_size;
     uint8_t bucket_shift;
     bool shared_allocator;
     int32_t bucket_count;
@@ -1639,25 +1640,28 @@ typedef struct ecs_map_iter_t {
     const ecs_map_t *map;
     ecs_bucket_t *bucket;
     ecs_bucket_entry_t *entry;
+    ecs_map_data_t *res;
 } ecs_map_iter_t;
 
 typedef struct ecs_map_params_t {
-    ecs_size_t size;
     struct ecs_allocator_t *allocator;
     struct ecs_block_allocator_t entry_allocator;
-    int32_t initial_count;
 } ecs_map_params_t;
 
 #define ECS_MAP_INIT(T) { .elem_size = ECS_SIZEOF(T) }
 
-FLECS_API
-void _ecs_map_params_init(
-    ecs_map_params_t *params,
-    struct ecs_allocator_t *allocator,
-    ecs_size_t elem_size);
+/* Function/macro postfixes meaning:
+ *   _ptr:    access ecs_map_val_t as void*
+ *   _ref:    access ecs_map_val_t* as T**
+ *   _deref:  dereferences a _ref
+ *   _alloc:  if _ptr is NULL, alloc
+ *   _free:   if _ptr is not NULL, free
+ */
 
-#define ecs_map_params_init(params, allocator, T)\
-    _ecs_map_params_init(params, allocator, ECS_SIZEOF(T))
+FLECS_API
+void ecs_map_params_init(
+    ecs_map_params_t *params,
+    struct ecs_allocator_t *allocator);
 
 FLECS_API
 void ecs_map_params_fini(
@@ -1665,110 +1669,80 @@ void ecs_map_params_fini(
 
 /** Initialize new map. */
 FLECS_API
-void _ecs_map_init(
+void ecs_map_init(
     ecs_map_t *map,
-    ecs_size_t elem_size,
-    struct ecs_allocator_t *allocator,
-    int32_t initial_count);
-
-#define ecs_map_init(map, T, allocator, initial_count)\
-    _ecs_map_init(map, ECS_SIZEOF(T), allocator, initial_count)
+    struct ecs_allocator_t *allocator);
 
 /** Initialize new map. */
 FLECS_API
-void _ecs_map_init_w_params(
+void ecs_map_init_w_params(
     ecs_map_t *map,
     ecs_map_params_t *params);
-
-#define ecs_map_init_w_params(map, param)\
-    _ecs_map_init_w_params(map, param)
 
 /** Initialize new map if uninitialized, leave as is otherwise */
 FLECS_API
-void _ecs_map_init_if(
+void ecs_map_init_if(
     ecs_map_t *map,
-    ecs_size_t elem_size,
-    struct ecs_allocator_t *allocator,
-    int32_t elem_count);
-
-#define ecs_map_init_if(map, T, allocator, elem_count)\
-    _ecs_map_init_if(map, ECS_SIZEOF(T), allocator, elem_count)
+    struct ecs_allocator_t *allocator);
 
 FLECS_API
-void _ecs_map_init_w_params_if(
+void ecs_map_init_w_params_if(
     ecs_map_t *result,
     ecs_map_params_t *params);
-
-#define ecs_map_init_w_params_if(map, params)\
-    _ecs_map_init_w_params_if(map, params)
 
 /** Deinitialize map. */
 FLECS_API
 void ecs_map_fini(
     ecs_map_t *map);
 
-/** Create new map. */
-FLECS_API
-ecs_map_t* _ecs_map_new(
-    ecs_size_t elem_size,
-    struct ecs_allocator_t *allocator,
-    int32_t elem_count);
-
-#define ecs_map_new(T, allocator, elem_count)\
-    _ecs_map_new(ECS_SIZEOF(T), allocator, elem_count)
-
 /** Get element for key, returns NULL if they key doesn't exist. */
 FLECS_API
-void* _ecs_map_get(
-    const ecs_map_t *map,
-    ecs_size_t elem_size,
-    ecs_map_key_t key);
-
-#define ecs_map_get(map, T, key)\
-    (T*)_ecs_map_get(map, ECS_SIZEOF(T), (ecs_map_key_t)key)
-
-/** Get pointer element. This dereferences the map element as a pointer. This
- * operation returns NULL when either the element does not exist or whether the
- * pointer is NULL, and should therefore only be used when the application knows
- * for sure that a pointer should never be NULL. */
-FLECS_API
-void* _ecs_map_get_ptr(
+ecs_map_val_t* ecs_map_get(
     const ecs_map_t *map,
     ecs_map_key_t key);
 
-#define ecs_map_get_ptr(map, T, key)\
-    (T)_ecs_map_get_ptr(map, key)
-
-/** Get or create element for key. */
+/* Get element as pointer (auto-dereferences _ptr) */
 FLECS_API
-void* _ecs_map_ensure(
+void* _ecs_map_get_deref(
+    const ecs_map_t *map,
+    ecs_map_key_t key);
+
+/** Get or insert element for key. */
+FLECS_API
+ecs_map_val_t* ecs_map_ensure(
+    ecs_map_t *map,
+    ecs_map_key_t key);
+
+/** Get or insert pointer element for key, allocate if the pointer is NULL */
+FLECS_API
+void* ecs_map_ensure_alloc(
     ecs_map_t *map,
     ecs_size_t elem_size,
     ecs_map_key_t key);
 
-#define ecs_map_ensure(map, T, key)\
-    ((T*)_ecs_map_ensure(map, ECS_SIZEOF(T), (ecs_map_key_t)key))
-
-/** Get or create element for key. */
+/** Insert element for key. */
 FLECS_API
-void* _ecs_map_insert(
+void ecs_map_insert(
+    ecs_map_t *map,
+    ecs_map_key_t key,
+    ecs_map_val_t value);
+
+/** Insert pointer element for key, populate with new allocation. */
+FLECS_API
+void* ecs_map_insert_alloc(
     ecs_map_t *map,
     ecs_size_t elem_size,
     ecs_map_key_t key);
 
-#define ecs_map_insert(map, T, key)\
-    ((T*)_ecs_map_insert(map, ECS_SIZEOF(T), (ecs_map_key_t)key))
-
-/** Free map. */
+/** Remove key from map. */
 FLECS_API
-void ecs_map_free(
-    ecs_map_t *map);
+ecs_map_val_t ecs_map_remove(
+    ecs_map_t *map,
+    ecs_map_key_t key);
 
-/** Remove key from map.
- * Returns number of remaining elements.
- */
+/* Remove pointer element, free if not NULL */
 FLECS_API
-int32_t ecs_map_remove(
+void ecs_map_remove_free(
     ecs_map_t *map,
     ecs_map_key_t key);
 
@@ -1790,41 +1764,29 @@ ecs_map_iter_t ecs_map_iter(
 
 /** Obtain next element in map from iterator. */
 FLECS_API
-void* _ecs_map_next(
-    ecs_map_iter_t* iter,
-    ecs_size_t elem_size,
-    ecs_map_key_t *key);
-
-#define ecs_map_next(iter, T, key) \
-    (T*)_ecs_map_next(iter, ECS_SIZEOF(T), key)
-
-/** Obtain next pointer element from iterator. See ecs_map_get_ptr. */
-FLECS_API
-void* _ecs_map_next_ptr(
-    ecs_map_iter_t* iter,
-    ecs_map_key_t *key);
-
-#define ecs_map_next_ptr(iter, T, key) \
-    (T)_ecs_map_next_ptr(iter, key)
+bool ecs_map_next(
+    ecs_map_iter_t *iter);
 
 /** Copy map. */
 FLECS_API
-ecs_map_t* ecs_map_copy(
-    ecs_map_t *map);
+void ecs_map_copy(
+    ecs_map_t *dst,
+    const ecs_map_t *src);
 
-#ifndef FLECS_LEGACY
-#define ecs_map_each(map, T, key, var, ...)\
-    {\
-        ecs_map_iter_t it = ecs_map_iter(map);\
-        ecs_map_key_t key;\
-        T* var;\
-        (void)key;\
-        (void)var;\
-        while ((var = ecs_map_next(&it, T, &key))) {\
-            __VA_ARGS__\
-        }\
-    }
-#endif
+#define ecs_map_get_ref(m, T, k) ECS_CAST(T**, ecs_map_get(m, k))
+#define ecs_map_get_deref(m, T, k) ECS_CAST(T*, _ecs_map_get_deref(m, k))
+#define ecs_map_ensure_ref(m, T, k) ECS_CAST(T**, ecs_map_ensure(m, k))
+#define ecs_map_ensure_deref(m, T, k) ECS_CAST(T*, ecs_map_ensure_ref(m, k)[0])
+
+#define ecs_map_insert_ptr(m, k, v) ecs_map_insert(m, k, ECS_CAST(ecs_map_val_t, v))
+#define ecs_map_insert_alloc_t(m, T, k) ECS_CAST(T*, ecs_map_insert_alloc(m, ECS_SIZEOF(T), k))
+#define ecs_map_ensure_alloc_t(m, T, k) ECS_CAST(T*, ecs_map_ensure_alloc(m, ECS_SIZEOF(T), k))
+#define ecs_map_remove_ptr(m, k) ((void*)(ecs_map_remove(m, k)))
+
+#define ecs_map_key(it) ((it)->res[0])
+#define ecs_map_value(it) ((it)->res[1])
+#define ecs_map_ptr(it) ECS_CAST(void*, ecs_map_value(it))
+#define ecs_map_ref(it, T) (ECS_CAST(T**, &((it)->res[1])))
 
 #ifdef __cplusplus
 }
@@ -3736,6 +3698,7 @@ typedef struct {
     ecs_size_t key_size;
     ecs_size_t value_size;
     ecs_block_allocator_t *hashmap_allocator;
+    ecs_block_allocator_t bucket_allocator;
     ecs_map_t impl;
 } ecs_hashmap_t;
 
@@ -3833,8 +3796,8 @@ void flecs_hm_bucket_remove(
 
 FLECS_DBG_API
 void flecs_hashmap_copy(
-    const ecs_hashmap_t *src,
-    ecs_hashmap_t *dst);
+    ecs_hashmap_t *dst,
+    const ecs_hashmap_t *src);
 
 FLECS_DBG_API
 flecs_hashmap_iter_t flecs_hashmap_iter(
@@ -12401,7 +12364,7 @@ typedef struct ecs_enum_constant_t {
 
 typedef struct EcsEnum {
     /** Populated from child entities with Constant component */
-    ecs_map_t *constants; /* map<i32_t, ecs_enum_constant_t> */
+    ecs_map_t constants; /* map<i32_t, ecs_enum_constant_t> */
 } EcsEnum;
 
 typedef struct ecs_bitmask_constant_t {
@@ -12417,7 +12380,7 @@ typedef struct ecs_bitmask_constant_t {
 
 typedef struct EcsBitmask {
     /* Populated from child entities with Constant component */
-    ecs_map_t *constants; /* map<u32_t, ecs_bitmask_constant_t> */
+    ecs_map_t constants; /* map<u32_t, ecs_bitmask_constant_t> */
 } EcsBitmask;
 
 typedef struct EcsArray {