--- a/include/flecs/private/block_allocator.h
+++ b/include/flecs/private/block_allocator.h
@@ -42,6 +42,11 @@ FLECS_API
 ecs_block_allocator_t* flecs_ballocator_new(
     ecs_size_t size);
 
+#define flecs_ballocator_new_t(T)\
+    flecs_ballocator_new(ECS_SIZEOF(T))
+#define flecs_ballocator_new_n(T, count)\
+    flecs_ballocator_new(ECS_SIZEOF(T) * count)
+
 FLECS_API
 void flecs_ballocator_fini(
     ecs_block_allocator_t *ba);