This PR contains the following improvements:
- A simpler (and slightly faster) map implementation
- A faster way to lookup id records with low ids

One benchmark where improvements are visible is the one for `ecs_get`:

|  Benchmark | Old | New  |
|------|------|------|
| get (empty entity) | 2.79ns | 2.96ns |
| get (id not found) | 7.07ns | 5.16ns |
| get (n ids) | 8.08ns | 6.37ns |
| ref_init | 6.51ns | 5.59ns |
| ref_get | 2.24ns | 2.24ns |